backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
fastapi[all]
python-dotenv>=0.19.0
supabase>=1.0.3
httpx>=0.26,<0.29
pydantic>=1.8.0
bcrypt>=3.2.0
PyJWT>=2.3.0
python-multipart>=0.0.5
google-generativeai>=0.3.0
sentence-transformers>=2.2.2

# Calendar specific requirements
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
httplib2>=0.20.0
PyPDF2>=3.0.0
pytesseract>=0.3.10

# Timezone support
pytz>=2023.3
